requirements.txt
Django==3.1.4
python-docx==0.8.10
django-bootstrap-modal-forms==2.0.1
psycopg2==2.8.6
django-heroku==0.3.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4